Adding requirements
